fix: Handle Int64 from compatible versions that do not share a prototype.
Added IInt64 interface and type check.
Fixes for Int64 created from a buffer and offset.
Fix negative check for toDecimalString() (bit operations on numbers from the buffer are 32-bit, not 8-bit).
Add tests and handling for NaN, Infinity, and -Infinity when setting the Int64 value from number(s).

Re-implement many of the number methods using Buffer number methods instead of byte by byte implementations.
Remove _2scomp() method that looks like it was intended to be private but ended up being used in the compact protocol.
Handle the -1 case in the compact protocol without using _2scomp().

Consolidate number read/writes for protocol/transport use.
Share a single toggle for asserting or not.
The code it replaced has no checks so not asserting is the closest equivalent but the checks might be preferred for safety.

Updated comments.